requirements.txt
azure-core==1.20.1
azure-identity==1.6.1
azure-kusto-data==2.3.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cryptography==35.0.0
idna==2.10
mo-future==3.147.20327
moz-sql-parser==3.32.20026
msal==1.16.0
msal-extensions==0.3.0
portalocker==1.7.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.3.1
python-dateutil==2.8.2
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.23
urllib3==1.26.3
